# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CSR Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: organizationId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/milestones:
    get:
      operationId: listProjectMilestones
      tags: [projects]
      summary: List project milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createProjectMilestone
      tags: [projects]
      summary: Add milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /projects/{id}/reports:
    get:
      operationId: listProjectReports
      tags: [projects]
      summary: List project reports
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectReport"
    post:
      operationId: createProjectReport
      tags: [projects]
      summary: Create project report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectReport"

  # CO-FUNDING
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        proposalId:
          type: integer
        title:
          type: string
        status:
          type: string
        organizationId:
          type: integer
        organizationName:
          type: string
        budgetTotal:
          type: integer
        budgetUsed:
          type: integer
        progressPercent:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        reports:
          type: array
          items:
            $ref: "#/components/schemas/ProjectReport"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - proposalId
        - title
        - status
        - organizationId
        - organizationName
        - budgetTotal
        - budgetUsed
        - progressPercent
        - milestones
        - reports
        - createdAt
        - updatedAt
    Project:
      type: object
      properties:
        id:
          type: integer
        proposalId:
          type: integer
        title:
          type: string
        status:
          type: string
        organizationId:
          type: integer
        organizationName:
          type: string
        budgetTotal:
          type: integer
        budgetUsed:
          type: integer
        progressPercent:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - proposalId
        - title
        - status
        - organizationId
        - organizationName
        - budgetTotal
        - budgetUsed
        - progressPercent
        - createdAt
        - updatedAt
    ProjectListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Milestone:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        status:
          type: string
        progressPercent:
          type: integer
      required:
        - id
        - projectId
        - title
        - status
        - progressPercent
    CreateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
    ProjectReport:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        content:
          type: string
        reportType:
          type: string
        budgetUsed:
          type: ["integer", "null"]
        beneficiariesCount:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - content
        - reportType
        - createdAt
    CreateProjectReportBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        reportType:
          type: string
        budgetUsed:
          type: ["integer", "null"]
        beneficiariesCount:
          type: ["integer", "null"]
      required:
        - title
        - content
        - reportType

    # CO-FUNDING
...